ECE-482 Testing video capture
Along with capybara_driver_base.rb, testing to see what caused Jenkins to not be able to obtain a stable Firefox connection.
The added if statement on line 70 will make lines 71-73 to run if it isn't Jenkins executing the code.

+++ features/support/env.rb
@@ -67,8 +67,10 @@ class KaikiWorld
   @@kaiki.record[:document_numbers] = ENV['KAIKI_DOC_NUMBERS'].split(',')     \
                                       if ENV['KAIKI_DOC_NUMBERS']
 
-  at_exit do
-    @@kaiki.headless.destroy if is_headless
+  if ENV['BUILD_NUMBER'].nil?
+    at_exit do
+      @@kaiki.headless.destroy if is_headless
+    end
   end
 
   def kaiki
